toggle fr tracking taken out from toolbar and made controlled by argument
