+++ stytra/gui/container_windows.py
@@ -1,7 +1,7 @@
 import logging
 import datetime
 
-from PyQt5.QtCore import Qt
+from PyQt5.QtCore import Qt, pyqtSignal
 from PyQt5.QtWidgets import (
     QLabel,
     QWidget,
@@ -108,6 +108,12 @@ class ExperimentWindow(QMainWindow):
         if experiment.trigger is not None:
             self.chk_scope = QCheckBox("Wait for trigger signal")
 
+        if experiment.motor_tracking is not None:
+            self.toggleMotor = ToggleIconButton(
+                icon_off="play", icon_on="stop", action_on="play", on=False
+            )
+            self.toggleMotor.clicked.connect(self.toggle_motor_tracking)
+
         self.logger = QPlainTextEditLogger()
         self.experiment.logger.addHandler(self.logger)
 
@@ -162,6 +168,9 @@ class ExperimentWindow(QMainWindow):
         if self.experiment.trigger is not None:
             self.toolbar_control.addWidget(self.chk_scope)
 
+        if self.experiment.motor_tracking is not None:
+            self.toolbar_control.addWidget(self.toggleMotor)
+
         self.experiment.gui_timer.timeout.connect(
                 self.plot_framerate.update)
 
@@ -176,6 +185,12 @@ class ExperimentWindow(QMainWindow):
             self.experiment.use_db = True
         else:
             self.experiment.use_db = False
+
+    def toggle_motor_tracking(self):
+        if self.toggleMotor.on == True:
+            self.toolbar_control.sig_start_tracking.emit()
+        else:
+            self.toolbar_control.sig_stop_tracking.emit()
 
     def closeEvent(self, *args, **kwargs):
         """
